README - initial commit
The point of this repo is to grow and explore SCSS architecture.

Goals: Establish a layered approach to styling that marries established SCSS practices such as 7-1 organization
and the use of @use with an awareness of component based web architecture while providing overriding capabilities
with 3rd party libraries.

For this repo initial commits will defer the application of 7-1 until complexity has risen to a level that it is warranted.

This repo will also take advantage of Angular as its web component framework of choice. Wijmo was selected as the
third party UI library with which to provide interop.